feat(scenarios): Create scenarios module to manage test scenarios
A Scenario is a Model, Spec and Options which can be run against any optimizer. Scenarios are more portable because a user can define a Scenario and it will be runnable with any optimizer so it is easier to export and move around. The staliro module was approaching this architecture so a separate module was created to separate the logic