refactor(specifications): Invert architecture so that results from models run their own evaluation
Instead of branching based on the type of the result of a simulation, enable the result types to evaluate themselves given a specification. By structuring this way, the generic data attribute type is not erased by isinstance checks and the logic that dictates the falsification returns a -inf is right in the implementation of the Falsification class.