refactor: Update class names, remove SystemInputs data class, and parameterize model data state data type
Remove sample decomposition logic present in SystemInputs class, which is not related to model simulation. Rename data classes from SystemData to ModelData to clearly indicate the relationship of the data classes. Parameterize ModelData data class state data type to enable models to indicate the shape of their output data. This works in tandem with the Specification state type parameterization to allow the type checker to verify that a model produces the right type for a specification to analyze