refactor: Update generic parameterizations to utilize new state type for both model instances and specification instances. Implement sample decomposition logic
CostFn class is now parameterized by a type variable StateT which indicates the type of the state data returned by the model and the type of state data the specification can handle. Since this type variable is shared, it ensures that the type checker will complain if they do not match. Sample decomposition logic is also moved to this module so that the model implementation does not need to know how to split the sample values