feat: Parameterize specifications based on state data type they can process. Add import checking for parser module to enable/disable TLTK specification. Update the predicate column map to be an int rather than a named tuple.
The predicate column map dtype field was often unused so it was removed. Feature checking was moved to this module to keep all the feature-based logic centralized. In addition, the error is closer to where users will use the functionality, so it makes sense that the error should be discoverable through code.