refactor: Move imports out of parse function body
The import behavior of the parse module is now full import or an error. As a result, the imports that were previously in the function body can now be moved to the module level. Consumers of the parser module can check if parsing is enabled by attempting to import the module and checking for any errors.