feat: Create layout module with SampleLayout class definition
A SampleLayout instance defines how a sample should be decomposed into static
parameters and signals. The interval of values that should be used as either the
static parameters or to generate a signal are represented as a tuple of integers
that indicate the starting and stopping index of the interval. By creating a
layout class like this it can be passed around as a single value instead of
needing to keep track of both the static parameters and signal slices
separately.