fix: Remove Attribute type from validate_signals function signature
Mypy complains that the Attribute class is missing a type annotation, but when
running the test suite an error is thrown because the Attribute class is not
subscriptable. Since the attr parameter is not used in the validator, it is
safe to make the parameter of Any type to resolve the issue.