refactor: Have users return a Mapping from Ode function not a Sequence
The state is provided to users in an ODE function as a dictionary, so it would make sense that the
derivatives should also be a dictionary so users don't need to guess the order of the variables. Since
SciPy requires that we use arrays, we create a variable order that we use to create arrays with a
consistent ordering.